requirements.txt
langchain
chromadb
fastapi
sentence-transformers
python-multipart
pypdf
ollama
PyPDF2
simsimd
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0